moved create project to page, added roles and edit support
